Factor render helpers out.
